+++ src/e2e/package.json
@@ -18,6 +18,7 @@
     "django": "./bin/backend-manage.sh",
     "db:flush": "npm run django flush -- --no-input",
     "db:bootstrap": "npm run django e2e_demo",
+    "db:bootstrap-clientbridge": "npm run django e2e_clientbridge",
     "db:reset": "npm run db:flush && npm run db:bootstrap"
   },
   "keywords": ["e2e", "playwright", "testing"],
