+++ src/e2e/src/__tests__/signatures.spec.ts
@@ -30,7 +30,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to signatures page
     const settingsButton = page.getByRole("button", { name: "More options" });
     await settingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // Wait for the data grid to load
@@ -92,7 +92,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to signatures page
     const settingsButton = page.getByRole("button", { name: "More options" });
     await settingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // First create a signature to edit
@@ -136,7 +136,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to signatures page
     const settingsButton = page.getByRole("button", { name: "More options" });
     await settingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // First create a signature to delete
@@ -174,7 +174,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to signatures page
     const settingsButton = page.getByRole("button", { name: "More options" });
     await settingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // Create a signature
@@ -208,7 +208,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to signatures page
     const settingsButton = page.getByRole("button", { name: "More options" });
     await settingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // Create a signature with default checkbox checked
@@ -237,7 +237,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to signatures page
     const settingsButton = page.getByRole("button", { name: "More options" });
     await settingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // Create a default signature
@@ -272,7 +272,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to signatures page
     const settingsButton = page.getByRole("button", { name: "More options" });
     await settingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // Create first signature and set as default
@@ -319,7 +319,7 @@ test.describe("Mailbox Signatures", () => {
     // Navigate to mailbox signatures page
     const userSettingsButton = page.getByRole("button", { name: "More options" });
     await userSettingsButton.click();
-    await page.getByRole("menuitem", { name: "Signatures" }).click();
+    await page.getByRole("menuitem", { name: "My signatures" }).click();
     await page.waitForURL("**/mailbox/*/signatures");
 
     // Create a mailbox signature WITH is_default
